Updated the classes for proper review
