+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scenes/StateSynchronization.unity
@@ -788,7 +788,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 118d29b9f26383b429c376bfa8e5bc7b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  debugLogging: 1
+  debugLogging: 0
   arucoMarkerDetector: {fileID: 1399172232}
 --- !u!1 &1869681819
 GameObject:
@@ -941,7 +941,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 3f820ead9dbb8f244b5614cf48d6247d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  debugLogging: 1
+  debugLogging: 0
   connectionManager: {fileID: 621974541}
   port: 7410
   anchorPrefab: {fileID: 1370450866519632, guid: 626257b1a6cd47c2a32a18cf75b2fb23,
@@ -990,10 +990,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6e82979a845c4c8469ffdce5ca285d84, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  debugLogging: 0
   connectionManager: {fileID: 621974541}
   anchorPrefab: {fileID: 1370450866519632, guid: 626257b1a6cd47c2a32a18cf75b2fb23,
     type: 3}
-  debugLogging: 1
   Port: 7410
 --- !u!4 &2075391330
 Transform:
