+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/StateSynchronization/Demo/ConnectedObserver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7645f4d2e21bf654e9090c1e884f17d0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
