+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/StateSynchronization/SpatialLocalization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 885d7e4848e31724089823f5b31cfc72
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
