+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/StateSynchronization/SpatialLocalization/HostCoordinateLocalizationMechanism.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6c5ed5c1752dc2f41888267ef35bb498
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
