+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/StateSynchronization/SpatialLocalization/LocalizationMechanismBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60489eeddde42474c8990b66af807f76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
